requirements.txt
langchain
faiss-cpu
sentence-transformers
streamlit
ollama
pypdf
python-docx
